# openapi.yaml
openapi: 3.0.3
info:
  title: AgentX Registry API
  version: 0.8.0
  description: |
    Hosted AgentX API. Business routes require a bearer token and return a stable
    ErrorEnvelope on JSON errors. Unknown dependency failures are logged with the
    request ID and return stable non-sensitive messages. Workspace roles are
    hierarchical: viewer, developer, admin, owner. Legacy unscoped routes are
    disabled in hosted mode. Device and release creation require idempotency keys.
    Expected duplicate and missing-resource outcomes use stable public 4xx envelopes;
    unexpected repository and dependency failures remain non-sensitive 500 responses.
security:
  - bearerAuth: []
servers:
  - url: https://{host}
    description: Hosted deployment
    variables:
      host: { default: registry.example.com }
paths:
...
  /v1/me:
    get:
      operationId: getCurrentPrincipal
      summary: Current authenticated principal
      responses:
        '200':
          description: Principal
          content:
            application/json: { schema: { '$ref': '#/components/schemas/Principal' } }
        '401': { '$ref': '#/components/responses/Unauthorized' }
    delete:
      operationId: deleteCurrentAccount
      summary: Delete the current account
      description: Atomically removes the user and non-owner memberships, deletes invitations addressed to the account email, anonymizes retained actor references, and emits an anonymous deletion audit/outbox record. The request fails while the account owns any Workspace or is under an active legal hold.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmation]
              properties:
                confirmation: { type: string, enum: [DELETE] }
      responses:
        '204': { description: Account data deleted }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '401': { '$ref': '#/components/responses/Unauthorized' }
        '500': { '$ref': '#/components/responses/InternalError' }
  /v1/me/export:
    get:
      operationId: exportCurrentAccount
      summary: Export the current account data
      description: Returns identity data stored for the current account, current Workspace memberships, invitations addressed to its verified email, and audit-event summaries attributed to it. Metadata belonging to a Workspace is not included.
      responses:
        '200':
          description: Complete current account export
          headers:
            Cache-Control: { schema: { type: string, enum: [no-store] } }
            Content-Disposition: { schema: { type: string } }
          content:
            application/json: { schema: { '$ref': '#/components/schemas/AccountExport' } }
        '401': { '$ref': '#/components/responses/Unauthorized' }
        '500': { '$ref': '#/components/responses/InternalError' }
  /v1/admin/legal-holds:
    get:
      operationId: listLegalHolds
      summary: List platform legal holds
      description: Requires a principal explicitly configured as a platform compliance administrator. Workspace roles, including owner, do not grant access.
      parameters:
        - { '$ref': '#/components/parameters/Limit' }
        - { '$ref': '#/components/parameters/Cursor' }
        - { name: status, in: query, required: false, schema: { type: string, enum: [active, released] } }
        - { name: target_type, in: query, required: false, schema: { '$ref': '#/components/schemas/LegalHoldTargetType' } }
        - { name: target_id, in: query, required: false, schema: { type: string } }
      responses:
        '200': { '$ref': '#/components/responses/LegalHoldPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '500': { '$ref': '#/components/responses/InternalError' }
    post:
      operationId: createLegalHold
      summary: Place a platform legal hold
      description: Requires a configured platform compliance administrator. An active hold blocks deletion of the target account or Workspace and excludes its attributed audit events from retention pruning.
      requestBody:
        required: true
        content:
          application/json: { schema: { '$ref': '#/components/schemas/LegalHoldCreate' } }
      responses:
        '201': { description: Legal hold created, content: { application/json: { schema: { '$ref': '#/components/schemas/LegalHold' } } } }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
        '500': { '$ref': '#/components/responses/InternalError' }
  /v1/admin/legal-holds/{id}/release:
    post:
      operationId: releaseLegalHold
      summary: Release a platform legal hold
      description: Requires a configured platform compliance administrator, the exact RELEASE confirmation, and a release reason. Released hold records and their lifecycle audits are retained.
      parameters: [{ '$ref': '#/components/parameters/LegalHoldID' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { '$ref': '#/components/schemas/LegalHoldRelease' } }
      responses:
        '200': { description: Legal hold released, content: { application/json: { schema: { '$ref': '#/components/schemas/LegalHold' } } } }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
        '500': { '$ref': '#/components/responses/InternalError' }
...
  /v1/workspaces/{id}/devices:
    get:
      operationId: listWorkspaceDevices
      summary: List workspace devices
      description: Requires viewer.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/DevicePage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
    post:
      operationId: registerWorkspaceDevice
      summary: Register or replace a device summary
      description: Requires developer. Device IDs are opaque text and unique only within the workspace. Retries with the same Idempotency-Key and request return the original response without duplicate audit or outbox events.
      x-agentx-minimum-role: developer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { '$ref': '#/components/schemas/Device' } }
      responses:
        '201':
          description: Registered device, or replayed with X-Idempotent-Replay true
          headers:
            X-Idempotent-Replay: { schema: { type: string, enum: ['true'] } }
          content:
            application/json: { schema: { '$ref': '#/components/schemas/Device' } }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
  /v1/workspaces/{id}/devices/{device_id}/heartbeat:
    post:
      operationId: heartbeatWorkspaceDevice
      summary: Update device and installed-package state
      description: Requires developer. device_id is workspace-scoped opaque text.
      x-agentx-minimum-role: developer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/DeviceID' }]
      requestBody:
        required: false
        content:
          application/json: { schema: { '$ref': '#/components/schemas/Device' } }
      responses:
        '204': { description: Heartbeat accepted }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
...
  /v1/workspaces/{id}/drift:
    get:
      operationId: listWorkspaceDrift
      summary: Compare devices with the current manifest
      description: Requires viewer.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/DriftPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
  /v1/workspaces/{id}/audit-events:
    get:
      operationId: listWorkspaceAuditEvents
      summary: List retention-managed audit events
      description: Requires viewer.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/AuditPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
...
  /v1/workspaces/{id}/manifest:
    get:
      operationId: getWorkspaceManifest
      summary: Get current desired-state manifest
      description: Requires viewer.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }]
      responses:
        '200': { description: Current manifest, content: { application/json: { schema: { '$ref': '#/components/schemas/TeamManifest' } } } }
        '403': { '$ref': '#/components/responses/Forbidden' }
    put:
      operationId: updateWorkspaceManifest
      summary: Replace desired-state manifest
      description: Requires admin. Package entries contain name, version, and sha256.
      x-agentx-minimum-role: admin
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { '$ref': '#/components/schemas/TeamManifestMutation' } }
      responses:
        '200': { description: New manifest revision, content: { application/json: { schema: { '$ref': '#/components/schemas/TeamManifest' } } } }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
  /v1/devices:
    get:
      operationId: listLegacyDevices
      deprecated: true
      summary: List devices in local legacy mode
      description: Disabled in hosted mode.
      parameters: [{ '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/DevicePage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '500': { '$ref': '#/components/responses/InternalError' }
    post:
      operationId: registerLegacyDevice
      deprecated: true
      summary: Register a device in local legacy mode
      description: Disabled in hosted mode.
      requestBody:
        required: true
        content:
          application/json: { schema: { '$ref': '#/components/schemas/Device' } }
      responses:
        '201': { description: Registered }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '500': { '$ref': '#/components/responses/InternalError' }
  /v1/devices/{id}/heartbeat:
    post:
      operationId: heartbeatLegacyDevice
      deprecated: true
      summary: Heartbeat in local legacy mode
      description: Disabled in hosted mode.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Updated }
        '404': { '$ref': '#/components/responses/NotFound' }
        '500': { '$ref': '#/components/responses/InternalError' }
...
  /v1/audit-events:
    get:
      operationId: listLegacyAuditEvents
      deprecated: true
      summary: List audit events in local legacy mode
      description: Disabled in hosted mode.
      parameters: [{ '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/AuditPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '500': { '$ref': '#/components/responses/InternalError' }
  /v1/drift:
    get:
      operationId: listLegacyDrift
      deprecated: true
      summary: List drift in local legacy mode
      description: Disabled in hosted mode.
      parameters: [{ '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/DriftPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '500': { '$ref': '#/components/responses/InternalError' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: OIDC JWT }
  parameters:
    WorkspaceID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    DeviceID: { name: device_id, in: path, required: true, description: Workspace-scoped opaque text identifier, schema: { type: string, minLength: 1 } }
...
    Limit: { name: limit, in: query, required: false, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
    Cursor: { name: cursor, in: query, required: false, description: Opaque value from next_cursor or X-Next-Cursor, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 255 } }
...
    LegalHoldID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
...
    LegalHoldPage: { description: Legal hold page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/LegalHold' } } } }] } } } }
...
    DevicePage: { description: Device page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/Device' } } } }] } } } }
    AuditPage: { description: Audit page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/AuditEvent' } } } }] } } } }
    DriftPage: { description: Drift page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/DriftItem' } } } }] } } } }
    BadRequest: { description: Invalid request, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Missing or invalid bearer token, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    Forbidden: { description: Workspace membership or role denied, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    InternalError: { description: Request failed without exposing dependency details, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
...
    Page:
      type: object
      required: [items, count]
      properties:
        items: { type: array, items: {} }
        count: { type: integer, minimum: 0, description: Total matching items before page slicing }
        next_cursor: { type: string }
    Principal:
      type: object
      required: [id, subject]
      properties:
        id: { type: string }
        issuer: { type: string, format: uri }
        subject: { type: string }
        email: { type: string, format: email }
        email_verified: { type: boolean, description: True only when the token carried an explicit boolean true email_verified claim }
    User:
      type: object
      required: [id, issuer, subject, email_verified, created_at]
      properties:
        id: { type: string }
        issuer: { type: string }
        subject: { type: string }
        email: { type: string, format: email }
        email_verified: { type: boolean }
        created_at: { type: string, format: date-time }
    AccountMembership:
      type: object
      required: [workspace, role, created_at]
      properties:
        workspace: { '$ref': '#/components/schemas/Workspace' }
        role: { '$ref': '#/components/schemas/Role' }
        created_at: { type: string, format: date-time }
    AccountAuditEvent:
      type: object
      required: [action, at]
      properties:
        workspace_id: { type: string, format: uuid }
        action: { type: string }
        resource_type: { type: string }
        resource_id: { type: string }
        request_id: { type: string }
        at: { type: string, format: date-time }
    AccountExport:
      type: object
      required: [schema_version, exported_at, user, memberships, invitations, audit_events]
      properties:
        schema_version: { type: integer, enum: [1] }
        exported_at: { type: string, format: date-time }
        user: { '$ref': '#/components/schemas/User' }
        memberships: { type: array, items: { '$ref': '#/components/schemas/AccountMembership' } }
        invitations: { type: array, items: { '$ref': '#/components/schemas/WorkspaceInvitation' } }
        audit_events: { type: array, items: { '$ref': '#/components/schemas/AccountAuditEvent' } }
    LegalHoldTargetType: { type: string, enum: [account, workspace] }
    LegalHoldCreate:
      type: object
      additionalProperties: false
      required: [target_type, target_id, reason]
      properties:
        target_type: { '$ref': '#/components/schemas/LegalHoldTargetType' }
        target_id: { type: string, minLength: 1 }
        reason: { type: string, minLength: 1, maxLength: 2000 }
    LegalHoldRelease:
      type: object
      additionalProperties: false
      required: [confirmation, reason]
      properties:
        confirmation: { type: string, enum: [RELEASE] }
        reason: { type: string, minLength: 1, maxLength: 2000 }
    LegalHold:
      type: object
      required: [id, target_type, target_id, reason, created_at]
      properties:
        id: { type: string, format: uuid }
        target_type: { '$ref': '#/components/schemas/LegalHoldTargetType' }
        target_id: { type: string }
        reason: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        released_by: { type: string }
        released_at: { type: string, format: date-time }
        release_reason: { type: string }
    Workspace:
      type: object
      required: [id, slug, name, created_at]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        created_at: { type: string, format: date-time }
    Role: { type: string, enum: [viewer, developer, admin, owner] }
    AssignableRole: { type: string, enum: [viewer, developer, admin] }
...
    WorkspaceInvitation:
      type: object
      required: [id, workspace_id, email, role, created_by, created_at, expires_at, status]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        workspace_name: { type: string }
        workspace_slug: { type: string }
        email: { type: string, format: email }
        role: { '$ref': '#/components/schemas/AssignableRole' }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time }
        revoked_by: { type: string }
        accepted_at: { type: string, format: date-time }
        accepted_by: { type: string }
        status: { type: string, enum: [pending, accepted, revoked, expired] }
...
    Device:
      type: object
      required: [name]
      properties:
        id: { type: string, description: Workspace-scoped opaque text; generated when omitted }
        name: { type: string, minLength: 1 }
        agent: { type: string }
        status: { type: string }
        updated_at: { type: string, format: date-time, readOnly: true }
        installed_packages:
          type: object
          additionalProperties: { type: string, pattern: '^[a-f0-9]{64}$' }
    AuditEvent:
      type: object
      required: [action, at]
      properties:
        action: { type: string }
        actor_id: { type: string }
        device_id: { type: string }
        resource_type: { type: string }
        resource_id: { type: string }
        metadata: { type: object, additionalProperties: true }
        request_id: { type: string }
        at: { type: string, format: date-time }
    DriftItem:
      type: object
      required: [device_name, package, kind, expected_sha256]
      properties:
        device_id: { type: string }
        device_name: { type: string }
        package: { type: string }
        kind: { type: string, enum: [missing, changed, extra] }
        expected_sha256: { type: string }
        observed_sha256: { type: string }
...
    TeamManifest:
      type: object
      required: [id, workspace_id, revision, document, created_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        revision: { type: integer, format: int64 }
        document: { '$ref': '#/components/schemas/DesiredManifestDocument' }
        created_at: { type: string, format: date-time }
    DesiredPackage:
      type: object
      additionalProperties: false
      required: [name, version, sha256]
      properties:
        name: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$' }
        version: { type: string, pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)(?:-[0-9A-Za-z-]+(?:\.[0-9A-Za-z-]+)*)?(?:\+[0-9A-Za-z-]+(?:\.[0-9A-Za-z-]+)*)?$' }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
    DesiredManifestDocument:
      type: object
      additionalProperties: false
      required: [version, packages]
      properties:
        version: { type: integer, enum: [1] }
        packages:
          type: array
          items: { '$ref': '#/components/schemas/DesiredPackage' }
    TeamManifestMutation:
      oneOf:
        - { '$ref': '#/components/schemas/DesiredManifestDocument' }
        - type: object
          additionalProperties: false
          required: [document]
          properties:
            document: { '$ref': '#/components/schemas/DesiredManifestDocument' }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string, minLength: 1 }